feat(ts-sdk): expose ledger signer utility functions (#10533)
Ports commit
https://github.com/MystenLabs/ts-sdks/commit/0faabdca34e65a4439fe926275a7c936150e0309,
without the `hasPublicObject` field.
Closes #10457
