+++ .changeset/small-walls-arrive.md
@@ -1,0 +1,5 @@
+---
+'@iota/signers': minor
+---
+
+Expose ledger signing utilities for external consumers.
