+++ sdk/signers/src/ledger/index.ts
@@ -10,8 +10,11 @@ import { Ed25519PublicKey } from '@iota/iota-sdk/keypairs/ed25519';
 import { Transaction } from '@iota/iota-sdk/transactions';
 import { toB64 } from '@iota/iota-sdk/utils';
 
-import { IotaMoveObject } from './bcs.js';
 import { bcs } from '@iota/iota-sdk/bcs';
+import { getInputObjects } from './objects.js';
+
+export { IotaMoveObject } from './bcs.js';
+export { getInputObjects } from './objects.js';
 
 /**
  * Configuration options for initializing the LedgerSigner.
@@ -67,7 +70,10 @@ export class LedgerSigner extends Signer {
      * @returns The signed transaction bytes and signature.
      */
     override async signTransaction(bytes: Uint8Array): Promise<SignatureWithBytes> {
-        const transactionOptions = await this.#getClearSigningOptions(bytes).catch(() => ({
+        const transactionOptions = await getInputObjects(
+            Transaction.from(bytes),
+            this.#iotaClient,
+        ).catch(() => ({
             // Fail gracefully so network errors or serialization issues don't break transaction signing:
             bcsObjects: [],
         }));
@@ -134,56 +140,6 @@ export class LedgerSigner extends Signer {
             ledgerClient,
             iotaClient,
         });
-    }
-
-    async #getClearSigningOptions(transactionBytes: Uint8Array) {
-        const transaction = Transaction.from(transactionBytes);
-        const data = transaction.getData();
-
-        const gasObjectIds = data.gasData.payment?.map((object) => object.objectId) ?? [];
-        const inputObjectIds = data.inputs
-            .map((input) => {
-                return input.$kind === 'Object' && input.Object.$kind === 'ImmOrOwnedObject'
-                    ? input.Object.ImmOrOwnedObject.objectId
-                    : null;
-            })
-            .filter((objectId): objectId is string => !!objectId);
-
-        const objects = await this.#iotaClient.multiGetObjects({
-            ids: [...gasObjectIds, ...inputObjectIds],
-            options: {
-                showBcs: true,
-                showPreviousTransaction: true,
-                showStorageRebate: true,
-                showOwner: true,
-            },
-        });
-
-        // NOTE: We should probably get rid of this manual serialization logic in favor of using the
-        // already serialized object bytes from the GraphQL API once there is more mainstream support
-        // for it + we can enforce the transport type on the Iota client.
-        const bcsObjects = objects
-            .map((object) => {
-                if (object.error || !object.data || object.data.bcs?.dataType !== 'moveObject') {
-                    return null;
-                }
-
-                return IotaMoveObject.serialize({
-                    data: {
-                        MoveObject: {
-                            type: object.data.bcs.type,
-                            version: object.data.bcs.version,
-                            contents: object.data.bcs.bcsBytes,
-                        },
-                    },
-                    owner: object.data.owner!,
-                    previousTransaction: object.data.previousTransaction!,
-                    storageRebate: object.data.storageRebate!,
-                }).toBytes();
-            })
-            .filter((bcsBytes): bcsBytes is Uint8Array => !!bcsBytes);
-
-        return { bcsObjects };
     }
 
     /**
